Separate workspace feature selector
